Ran black formatter on files

+++ excel_cvs_conversion.py
@@ -2,51 +2,56 @@
 from dotenv import load_dotenv
 import pandas as pd
 import sqlalchemy
+
 # import csv
 import os
 
 COLS_TO_USE = [
-    'CASE_NUMBER',
-    'VISA_CLASS',
-    'JOB_TITLE',
-    'SOC_TITLE',
-    'FULL_TIME_POSITION',
-    'EMPLOYER_NAME',
-    'EMPLOYER_CITY',
-    'EMPLOYER_STATE',
-    'WAGE_RATE_OF_PAY_FROM',
-    'WAGE_RATE_OF_PAY_TO',
-    'WAGE_UNIT_OF_PAY',
-    'PREVAILING_WAGE',
-    'PW_UNIT_OF_PAY',
+    "CASE_NUMBER",
+    "VISA_CLASS",
+    "JOB_TITLE",
+    "SOC_TITLE",
+    "FULL_TIME_POSITION",
+    "EMPLOYER_NAME",
+    "EMPLOYER_CITY",
+    "EMPLOYER_STATE",
+    "WAGE_RATE_OF_PAY_FROM",
+    "WAGE_RATE_OF_PAY_TO",
+    "WAGE_UNIT_OF_PAY",
+    "PREVAILING_WAGE",
+    "PW_UNIT_OF_PAY",
 ]
-CSV_FILENAME = 'test'
+CSV_FILENAME = "test"
 
 
 def csv_to_df(csv_filename):
     # Convert excel to csv
     df = pd.read_excel(
-        f'data/{csv_filename}.xlsx',
+        f"data/{csv_filename}.xlsx",
         usecols=COLS_TO_USE,
     )
     df = coerce_to_numeric_type(df)
     df = strip_and_standardize_casing(df)
-    df['WAGE_RATE_OF_PAY_FROM'] = df['WAGE_RATE_OF_PAY_FROM'].astype(float)
-    df['WAGE_RATE_OF_PAY_TO'] = df['WAGE_RATE_OF_PAY_TO'].astype(float)
-    df['PREVAILING_WAGE'] = df['PREVAILING_WAGE'].astype(float)
+    df["WAGE_RATE_OF_PAY_FROM"] = df["WAGE_RATE_OF_PAY_FROM"].astype(float)
+    df["WAGE_RATE_OF_PAY_TO"] = df["WAGE_RATE_OF_PAY_TO"].astype(float)
+    df["PREVAILING_WAGE"] = df["PREVAILING_WAGE"].astype(float)
 
     return df
 
 
 def coerce_to_numeric_type(df):
-    #̶ T̶O̶D̶O̶:̶ C̶h̶a̶n̶g̶e̶ t̶o̶ l̶i̶s̶t̶ c̶o̶m̶p̶r̶e̶h̶e̶n̶s̶i̶o̶n̶
+    # ̶ T̶O̶D̶O̶:̶ C̶h̶a̶n̶g̶e̶ t̶o̶ l̶i̶s̶t̶ c̶o̶m̶p̶r̶e̶h̶e̶n̶s̶i̶o̶n̶
     # List comprehension below doesn't work because list comprehension turns this into a list and not an object as required to mutate an entire dataframe
     # df = [df.iloc[:, row].str.strip() for row in range(len(df.columns)) if df.iloc[:, row].dtype == 'object']
 
     # Converting float column with strings to turn them into NaN. Otherwise, stripping will get rid of numbers in columns with strings
-    df['PREVAILING_WAGE'] = pd.to_numeric(df['PREVAILING_WAGE'], errors='coerce')
-    df['WAGE_RATE_OF_PAY_FROM'] = pd.to_numeric(df['WAGE_RATE_OF_PAY_FROM'], errors='coerce')
-    df['WAGE_RATE_OF_PAY_TO'] = pd.to_numeric(df['WAGE_RATE_OF_PAY_TO'], errors='coerce')
+    df["PREVAILING_WAGE"] = pd.to_numeric(df["PREVAILING_WAGE"], errors="coerce")
+    df["WAGE_RATE_OF_PAY_FROM"] = pd.to_numeric(
+        df["WAGE_RATE_OF_PAY_FROM"], errors="coerce"
+    )
+    df["WAGE_RATE_OF_PAY_TO"] = pd.to_numeric(
+        df["WAGE_RATE_OF_PAY_TO"], errors="coerce"
+    )
 
     return df
 
@@ -54,11 +59,13 @@ def coerce_to_numeric_type(df):
 def strip_and_standardize_casing(df):
     # Get rid of all the spaces and standardize casing for strings
     # Create an object out of all the columns with the dtype object
-    df_object = df.select_dtypes(['object'])
+    df_object = df.select_dtypes(["object"])
 
     # Modify key values
-        # Strips away leading and trailing spaces
-    df[df_object.columns] = df_object.apply(lambda x: x.str.strip()).apply(lambda x: x.str.title())
+    # Strips away leading and trailing spaces
+    df[df_object.columns] = df_object.apply(lambda x: x.str.strip()).apply(
+        lambda x: x.str.title()
+    )
 
     return df
 
@@ -70,12 +77,10 @@ def connect_to_database():
     Return: Engine object, provides access to a Connection, which can then invoke SQL statements
     """
     # Export the dataframe to the database
-    connection_string = os.path.expandvars(
-      os.environ['CRDB_CONN_STR']
-    ).replace(
-        'postgres://', 'cockroachdb://'
-    ).replace(
-        'postgresql://', 'cockroachdb://'
+    connection_string = (
+        os.path.expandvars(os.environ["CRDB_CONN_STR"])
+        .replace("postgres://", "cockroachdb://")
+        .replace("postgresql://", "cockroachdb://")
     )
     engine = sqlalchemy.create_engine(connection_string)
     return engine
@@ -91,7 +96,7 @@ def export_dataframe_to_database(df, engine, table_name):
     Return: N/A
     """
     print("Exporting dataframe to database...")
-    df.to_sql(f'{table_name}', engine, if_exists='replace', index=False)
+    df.to_sql(f"{table_name}", engine, if_exists="replace", index=False)
     print("Dataframe exported to database successfully!")
 
 
@@ -102,15 +107,21 @@ def create_small_dataframe():
     Reference: https://www.geeksforgeeks.org/applying-lambda-functions-to-pandas-dataframe/
     TO BE REMOVED LATER
     """
-    values= [['Rohan',455],['Elvish',250],['Deepak',495],
-            ['Soni',400],['Radhika',350],['Vansh',450]]
+    values = [
+        ["Rohan", 455],
+        ["Elvish", 250],
+        ["Deepak", 495],
+        ["Soni", 400],
+        ["Radhika", 350],
+        ["Vansh", 450],
+    ]
 
-    df = pd.DataFrame(values,columns=['Name','Total_Marks'])
-    df = df.assign(Percentage = lambda x: (x['Total_Marks'] /500 * 100))
+    df = pd.DataFrame(values, columns=["Name", "Total_Marks"])
+    df = df.assign(Percentage=lambda x: (x["Total_Marks"] / 500 * 100))
     return df
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
 
     # Get environment variables set in the .env file
     load_dotenv()
@@ -136,15 +147,13 @@ if __name__ == '__main__':
 
     # Print all rows in table
     with engine.connect() as conn:
-        result = conn.execute(
-            sqlalchemy.select([metadata.tables['exam_score']])
-        )
+        result = conn.execute(sqlalchemy.select([metadata.tables["exam_score"]]))
         for row in result:
             print(row)
 
     # Print df from db table
     print("Printing dataframe from db table...")
-    print(pd.read_sql('exam_score', engine))
+    print(pd.read_sql("exam_score", engine))
 
     # TODO: remove below function later
     # Clear all table stored in specified metadata
